+++ src/OpenTelemetry.Instrumentation.StackExchangeRedis/StackExchangeRedisCallsInstrumentation.cs
@@ -14,11 +14,8 @@
 // limitations under the License.
 // </copyright>
 
-using System;
 using System.Collections.Concurrent;
-using System.Collections.Generic;
 using System.Diagnostics;
-using System.Threading;
 using OpenTelemetry.Instrumentation.StackExchangeRedis.Implementation;
 using OpenTelemetry.Internal;
 using OpenTelemetry.Trace;
@@ -38,9 +35,9 @@ internal class StackExchangeRedisCallsInstrumentation : IDisposable
     internal static readonly string ActivityName = ActivitySourceName + ".Execute";
     internal static readonly Version Version = typeof(StackExchangeRedisCallsInstrumentation).Assembly.GetName().Version;
     internal static readonly ActivitySource ActivitySource = new(ActivitySourceName, Version.ToString());
-    internal static readonly IEnumerable<KeyValuePair<string, object>> CreationTags = new[]
+    internal static readonly IEnumerable<KeyValuePair<string, object?>> CreationTags = new[]
     {
-        new KeyValuePair<string, object>(SemanticConventions.AttributeDbSystem, "redis"),
+        new KeyValuePair<string, object?>(SemanticConventions.AttributeDbSystem, "redis"),
     };
 
     internal readonly ConcurrentDictionary<(ActivityTraceId TraceId, ActivitySpanId SpanId), (Activity Activity, ProfilingSession Session)> Cache
@@ -77,7 +74,7 @@ internal class StackExchangeRedisCallsInstrumentation : IDisposable
     /// Returns session for the Redis calls recording.
     /// </summary>
     /// <returns>Session associated with the current span context to record Redis calls.</returns>
-    public Func<ProfilingSession> GetProfilerSessionsFactory()
+    public Func<ProfilingSession?> GetProfilerSessionsFactory()
     {
         return () =>
         {
@@ -86,7 +83,7 @@ internal class StackExchangeRedisCallsInstrumentation : IDisposable
                 return null;
             }
 
-            Activity parent = Activity.Current;
+            var parent = Activity.Current;
 
             // If no parent use the default session.
             if (parent == null || parent.IdFormat != ActivityIdFormat.W3C)