+++ src/OpenTelemetry.Instrumentation.StackExchangeRedis/StackExchangeRedisCallsInstrumentationOptions.cs
@@ -14,7 +14,6 @@
 // limitations under the License.
 // </copyright>
 
-using System;
 using System.Diagnostics;
 using OpenTelemetry.Trace;
 using StackExchange.Redis.Profiling;
@@ -43,7 +42,7 @@ public class StackExchangeRedisCallsInstrumentationOptions
     /// <para><see cref="Activity"/>: the activity being enriched.</para>
     /// <para><see cref="IProfiledCommand"/>: the profiled redis command from which additional information can be extracted to enrich the activity.</para>
     /// </remarks>
-    public Action<Activity, IProfiledCommand> Enrich { get; set; }
+    public Action<Activity, IProfiledCommand>? Enrich { get; set; }
 
     /// <summary>
     /// Gets or sets a value indicating whether or not the <see cref="StackExchangeRedisCallsInstrumentation"/> should enrich Activity with <see cref="ActivityEvent"/> entries about the Redis command processing/lifetime. Defaults to <see cref="bool">true</see>.