+++ test/OpenTelemetry.Instrumentation.StackExchangeRedis.Tests/Implementation/RedisProfilerEntryToActivityConverterTests.cs
@@ -63,7 +63,7 @@ public class RedisProfilerEntryToActivityConverterTests : IDisposable
         profiledCommand.Setup(m => m.CommandCreated).Returns(DateTime.UtcNow);
         profiledCommand.Setup(m => m.Command).Returns("SET");
 
-        var result = RedisProfilerEntryToActivityConverter.ProfilerCommandToActivity(activity, profiledCommand.Object, new StackExchangeRedisCallsInstrumentationOptions());
+        var result = RedisProfilerEntryToActivityConverter.ProfilerCommandToActivity(activity, profiledCommand.Object, new StackExchangeRedisInstrumentationOptions());
 
         Assert.Equal("SET", result.DisplayName);
     }
@@ -76,7 +76,7 @@ public class RedisProfilerEntryToActivityConverterTests : IDisposable
         var profiledCommand = new Mock<IProfiledCommand>();
         profiledCommand.Setup(m => m.CommandCreated).Returns(now.DateTime);
 
-        var result = RedisProfilerEntryToActivityConverter.ProfilerCommandToActivity(activity, profiledCommand.Object, new StackExchangeRedisCallsInstrumentationOptions());
+        var result = RedisProfilerEntryToActivityConverter.ProfilerCommandToActivity(activity, profiledCommand.Object, new StackExchangeRedisInstrumentationOptions());
 
         Assert.Equal(now, result.StartTimeUtc);
     }
@@ -88,7 +88,7 @@ public class RedisProfilerEntryToActivityConverterTests : IDisposable
         var profiledCommand = new Mock<IProfiledCommand>();
         profiledCommand.Setup(m => m.CommandCreated).Returns(DateTime.UtcNow);
 
-        var result = RedisProfilerEntryToActivityConverter.ProfilerCommandToActivity(activity, profiledCommand.Object, new StackExchangeRedisCallsInstrumentationOptions());
+        var result = RedisProfilerEntryToActivityConverter.ProfilerCommandToActivity(activity, profiledCommand.Object, new StackExchangeRedisInstrumentationOptions());
 
         Assert.NotNull(result.GetTagValue(SemanticConventions.AttributeDbSystem));
         Assert.Equal("redis", result.GetTagValue(SemanticConventions.AttributeDbSystem));
@@ -102,7 +102,7 @@ public class RedisProfilerEntryToActivityConverterTests : IDisposable
         profiledCommand.Setup(m => m.CommandCreated).Returns(DateTime.UtcNow);
         profiledCommand.Setup(m => m.Command).Returns("SET");
 
-        var result = RedisProfilerEntryToActivityConverter.ProfilerCommandToActivity(activity, profiledCommand.Object, new StackExchangeRedisCallsInstrumentationOptions());
+        var result = RedisProfilerEntryToActivityConverter.ProfilerCommandToActivity(activity, profiledCommand.Object, new StackExchangeRedisInstrumentationOptions());
 
         Assert.NotNull(result.GetTagValue(SemanticConventions.AttributeDbStatement));
         Assert.Equal("SET", result.GetTagValue(SemanticConventions.AttributeDbStatement));
@@ -118,10 +118,10 @@ public class RedisProfilerEntryToActivityConverterTests : IDisposable
                             CommandFlags.NoRedirect;
         profiledCommand.Setup(m => m.Flags).Returns(expectedFlags);
 
-        var result = RedisProfilerEntryToActivityConverter.ProfilerCommandToActivity(activity, profiledCommand.Object, new StackExchangeRedisCallsInstrumentationOptions());
+        var result = RedisProfilerEntryToActivityConverter.ProfilerCommandToActivity(activity, profiledCommand.Object, new StackExchangeRedisInstrumentationOptions());
 
-        Assert.NotNull(result.GetTagValue(StackExchangeRedisCallsInstrumentation.RedisFlagsKeyName));
-        Assert.Equal("PreferMaster, FireAndForget, NoRedirect", result.GetTagValue(StackExchangeRedisCallsInstrumentation.RedisFlagsKeyName));
+        Assert.NotNull(result.GetTagValue(StackExchangeRedisConnectionInstrumentation.RedisFlagsKeyName));
+        Assert.Equal("PreferMaster, FireAndForget, NoRedirect", result.GetTagValue(StackExchangeRedisConnectionInstrumentation.RedisFlagsKeyName));
     }
 
     [Fact]
@@ -136,7 +136,7 @@ public class RedisProfilerEntryToActivityConverterTests : IDisposable
         profiledCommand.Setup(m => m.CommandCreated).Returns(DateTime.UtcNow);
         profiledCommand.Setup(m => m.EndPoint).Returns(ipLocalEndPoint);
 
-        var result = RedisProfilerEntryToActivityConverter.ProfilerCommandToActivity(activity, profiledCommand.Object, new StackExchangeRedisCallsInstrumentationOptions());
+        var result = RedisProfilerEntryToActivityConverter.ProfilerCommandToActivity(activity, profiledCommand.Object, new StackExchangeRedisInstrumentationOptions());
 
         Assert.NotNull(result.GetTagValue(SemanticConventions.AttributeNetPeerIp));
         Assert.Equal($"{address}.0.0.0", result.GetTagValue(SemanticConventions.AttributeNetPeerIp));
@@ -154,7 +154,7 @@ public class RedisProfilerEntryToActivityConverterTests : IDisposable
         profiledCommand.Setup(m => m.CommandCreated).Returns(DateTime.UtcNow);
         profiledCommand.Setup(m => m.EndPoint).Returns(dnsEndPoint);
 
-        var result = RedisProfilerEntryToActivityConverter.ProfilerCommandToActivity(activity, profiledCommand.Object, new StackExchangeRedisCallsInstrumentationOptions());
+        var result = RedisProfilerEntryToActivityConverter.ProfilerCommandToActivity(activity, profiledCommand.Object, new StackExchangeRedisInstrumentationOptions());
 
         Assert.NotNull(result.GetTagValue(SemanticConventions.AttributeNetPeerName));
         Assert.Equal(dnsEndPoint.Host, result.GetTagValue(SemanticConventions.AttributeNetPeerName));
@@ -172,7 +172,7 @@ public class RedisProfilerEntryToActivityConverterTests : IDisposable
         profiledCommand.Setup(m => m.CommandCreated).Returns(DateTime.UtcNow);
         profiledCommand.Setup(m => m.EndPoint).Returns(unixEndPoint);
 
-        var result = RedisProfilerEntryToActivityConverter.ProfilerCommandToActivity(activity, profiledCommand.Object, new StackExchangeRedisCallsInstrumentationOptions());
+        var result = RedisProfilerEntryToActivityConverter.ProfilerCommandToActivity(activity, profiledCommand.Object, new StackExchangeRedisInstrumentationOptions());
 
         Assert.NotNull(result.GetTagValue(SemanticConventions.AttributePeerService));
         Assert.Equal(unixEndPoint.ToString(), result.GetTagValue(SemanticConventions.AttributePeerService));